reduce results list with accumulator function
Instead of using a loop in which the outside variable `result` is
repeatedly modified, reduce the results list directly.

The semantics are slightly different.  The previous version would reset
`result` to the current item in the results list whenever `result` was
previously set to `None`.  `None` is not an expected output of
`_cleanup_intersect_gff_gff` so this would never have happened.  Hence,
the reduce version has the same behaviour.