Move initialisation to Genome and Regulator classes.
We don't want to store the dictionaries for all genomes and all
regulators in utils.py but rather in the relevant classes.  This is a
bit ugly as we need to write to class state once on initialisation.  The
upside is that we no longer need to pass around "dorina.utils.genome" or
"dorina.utils.regulators" to simple static methods like "from_name" or
"path_by_name".